moves refactored files to login/ where all new components are kept
